Changes to levelfactory. Extended the "grid". Mapping of number to entities and send the coordinates. Loop level and create (push) each entity within the mapping.
Changes to stars
Refactored stars to recive and handle the coordinates from levelfactory.
Removed unused code

Changes to gameboard
Added function createEntitiesForLevel with current level1.
Commented out new Star() for current test of levelfactory.
That whole section is to be removed since we will create/initialize the enitites in levelfactory.

Co-authored-by: gitwilma <wilma.haakansson@gmail.com>